# rpc: add BenchmarkGRPCPing